rift/pharma_guard/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.1
python-multipart>=0.0.12
pydantic>=2.9.0